+++ simurg_plotter/__init__.py
@@ -1,0 +1,1 @@
+from .global_ionospheric_map import visualization
